refactor(main): move the browser and file routes out of #handle
Two more modules, 13 routes. The moved text is character-identical to what it replaced once
`this.#json(response, ...)` becomes `json(...)` and `this.#options.x` becomes a destructured
dependency; checked mechanically, not by eye.

`route-files.ts` holds the four routes that write a body themselves rather than through `json`,
which makes it the first place where the `return "handled"` rule earns its keep: a bare `return`
would read as `"unmatched"` and let the router answer 404 over a response it had already ended.

Checked: `bun run test:desktop -- src/main/team-api-server.test.ts` (39 passed),
`typecheck:node`, `biome check`.
